pyproject.toml
[project]
name = "ai-soc-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-generativeai>=0.8.6",
    "groq>=1.1.2",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.13.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.38",
    "aiosqlite>=0.21.0",
    "bcrypt<4.0.0",
    "uvicorn[standard]>=0.44.0",
    "websockets>=16.0",
]

